[REFACTOR][I] Move VirtualFileConverter to package filesystem
Moves VirtualFileConverter from the package editor to package the
filesystem as it deals with resources and not editors.
